requirements.txt
# TIGA Hunt — dependencies
# All Apache 2.0 or MIT licensed

# API + UI
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
streamlit>=1.35.0
httpx>=0.27.0

# Vector store
chromadb>=0.5.0

# Ollama client
ollama>=0.2.0

# Document extraction
pypdf>=4.2.0
python-docx>=1.1.0
python-pptx>=0.6.23

# OCR (opt-in only — requires Tesseract binary installed separately)
pytesseract>=0.3.13
Pillow>=10.3.0

# Config
PyYAML>=6.0.1

# Utilities
tqdm>=4.66.0

# Testing
pytest>=8.0.0
